requirements.txt
# Core Dependencies
streamlit==1.28.1
pandas==2.1.3
numpy==1.25.2
plotly==5.17.0

# Data Processing
pyspark==3.4.1
delta-spark==2.4.0
sqlalchemy==2.0.23

# AI/ML Dependencies  
openai==1.3.7
langchain==0.0.350
tiktoken==0.5.1
sentence-transformers==2.2.2
scikit-learn==1.3.2

# Analytics and Visualization
seaborn==0.12.2
matplotlib==3.8.2
scipy==1.11.4
ydata-profiling==4.6.4

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.4.2
pyyaml==6.0.1

# (Portfolio demo) Development/production extras removed to reduce footprint
